+++ aed1/lista/src/lista.c
@@ -37,7 +37,7 @@ void append_values_l(Int_Linked_List* l, int n, int values[]) {
     }
 }
 
-void append_values_reverse_l(Int_Linked_List* l, int n, long values[]) {
+void append_values_reverse_l(Int_Linked_List* l, int n, int values[]) {
     int j;
     for(j = n - 1; j >= 0; j--) {
         append_l(l, values[j]);
